requirements.txt
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Marshmallow==0.14.0
Flask-JWT-Extended==4.3.1
marshmallow-sqlalchemy==0.26.1
psycopg2-binary==2.9.1
python-dotenv==0.19.0